requirements.txt
flask
pyyaml
flask_cors
sqlalchemy
pymysql
cryptography
pydantic